feat: create BackButtonComponent
